feat: Added color variable support to parse from YAML
Implemented the parsing support for the color YAML field, can parse the
given implementations as described in the
[docs](https://esphome.io/components/display/index.html#color).

Right now the white value is dropped from the color, because this seems
to be part of RGBW spec for LED lights, not necessarily for displays.
Might have to change this later if this is actually used as
opacity/alpha value.